+++ src/app/(route)/mypage/_components/Pagination.tsx
@@ -5,7 +5,7 @@ import { PageInfo } from "../_types/toolbarType";
 import Pg_left from "@/app/_components/icon/Pg_left";
 import Pg_right from "@/app/_components/icon/Pg_right";
 import Pg_double_right from "@/app/_components/icon/Pg_double_right";
-import { useEffect, useState } from "react";
+import useIsMobile from "@/utils/useIsMobile";
 
 interface PaginationProps {
   pageInfo: PageInfo;
@@ -17,18 +17,7 @@ const Pagination = ({ pageInfo, onPageChangeAction }: PaginationProps) => {
     currentPage: 1,
     totalPage: 1,
   };
-  const [isMobile, setIsMobile] = useState(false);
-
-  useEffect(() => {
-    const handleResize = () => {
-      setIsMobile(window.innerWidth < 768);
-    };
-
-    handleResize();
-
-    window.addEventListener("resize", handleResize);
-    return () => window.removeEventListener("resize", handleResize);
-  }, []);
+  const isMobile = useIsMobile();
 
   const getPageNumbers = () => {
     const maxVisiblePage = isMobile ? 4 : 5;
